Refactor Program.cs to extract methods
BingBacService now makes much more sense and handles all of the
JSON related functionality. It has methods to return values
from the JSON so that Program.cs doesn't need to know how this happens.

There are also two new Helpers that provide sensible wrapping of the
FileSystem and User32 SystemParameters (which sets the users
background).